@@ -1,11 +1,25 @@
 package com.ZharikovaES.PersonalListsApp.services;
 
-import lombok.Getter;
-import lombok.Setter;
-
-@Setter
-@Getter
 public class JwtRequest {
     private String login;
     private String password;
+
+    public JwtRequest() {
+    }
+
+    public String getLogin() {
+        return login;
+    }
+
+    public void setLogin(String login) {
+        this.login = login;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
