@@ -1,31 +1,34 @@
 package com.ZharikovaES.PersonalListsApp.services;
 
-import io.jsonwebtoken.Claims;
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
+import com.ZharikovaES.PersonalListsApp.models.Role;
+import com.auth0.jwt.interfaces.DecodedJWT;
 
+import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
-import java.util.stream.Collectors;
 
-import com.ZharikovaES.PersonalListsApp.models.Role;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
 public final class JwtUtils {
 
-    public static JwtAuthentication generate(Claims claims) {
-        final JwtAuthentication jwtInfoToken = new JwtAuthentication();
-        jwtInfoToken.setRoles(getRoles(claims));
-        jwtInfoToken.setFirstName(claims.get("firstName", String.class));
-        jwtInfoToken.setUsername(claims.getSubject());
+    private JwtUtils() {
+        // private constructor to prevent instantiation
+    }
+
+    public static JwtAuthentication generate(DecodedJWT jwt) {
+        JwtAuthentication jwtInfoToken = new JwtAuthentication();
+        jwtInfoToken.setUsername(jwt.getSubject());
+        jwtInfoToken.setFirstName(jwt.getClaim("firstName").asString());
+        jwtInfoToken.setRoles(getRoles(jwt));
         return jwtInfoToken;
     }
 
-    private static Set<Role> getRoles(Claims claims) {
-        final List<String> roles = claims.get("roles", List.class);
-        return roles.stream()
-                .map(Role::valueOf)
-                .collect(Collectors.toSet());
+    private static Set<Role> getRoles(DecodedJWT jwt) {
+        List<String> roles = jwt.getClaim("roles").asList(String.class);
+        Set<Role> roleSet = new HashSet<>();
+        if (roles != null) {
+            for (String role : roles) {
+                roleSet.add(Role.valueOf(role));
+            }
+        }
+        return roleSet;
     }
-
 }
