@@ -1,11 +1,24 @@
 package com.ZharikovaES.PersonalListsApp.services;
 
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
 public class LoginRequest {
-  private String username;
-  private String password;
+    private String username;
+    private String password;
+
+    public LoginRequest() {}
+
+    public String getUsername() {
+        return username;
+    }
+
+    public void setUsername(String username) {
+        this.username = username;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
 }
