@@ -1,12 +1,31 @@
 package com.ZharikovaES.PersonalListsApp.services;
 
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
 public class RegistrationRequest {
   private String username;
   private String password;
   private String email;
+
+  public String getUsername() {
+    return username;
+  }
+
+  public void setUsername(String username) {
+      this.username = username;
+  }
+
+  public String getPassword() {
+    return password;
+  }
+
+  public void setPassword(String password) {
+      this.password = password;
+  }
+
+  public String getEmail() {
+    return email;
+  }
+
+  public void setEmail(String email) {
+      this.email = email;
+  }
 }
