@@ -1,10 +1,17 @@
 package com.ZharikovaES.PersonalListsApp.services;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
 public class RegistrationResponse {
   private String message;
+
+  public RegistrationResponse(String message) {
+    this.message = message;
+  }
+
+  public String getMessage() {
+    return message;
+  }
+
+  public void setMessage(String message) {
+      this.message = message;
+  }
 }
